# cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LLMAppConfig:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      config_name:
        type: string
      llmapp_type:
        type: string
    type: object
  dto.Policy:
    properties:
      max_count:
        type: integer
      policyname:
        type: string
      repeat:
        type: integer
      week:
        type: string
      work_times:
        items:
          $ref: '#/definitions/dto.WorkTime'
        type: array
    type: object
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.WorkTime:
    properties:
      end_time:
        type: string
      start_time:
        type: string
    type: object
  request.KF:
    properties:
      bot_prompt:
        type: string
      bot_timeout:
        type: integer
      bot_timeout_msg:
        type: string
      bot_welcome_msg:
        type: string
      botid:
        type: string
      chat_timeout:
        type: integer
      chatend_msg:
        type: string
      kfid:
        type: string
      kfname:
        type: string
      kfplatform:
        type: string
      receive_priority:
        type: integer
      receive_rule:
        type: integer
      staff_list:
        items:
          type: string
        type: array
      staff_welcome_msg:
        type: string
      status:
        type: integer
      transfer_keywords:
        type: string
      unmanned_msg:
        type: string
    type: object
  request.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost
info:
  contact: {}
  description: evobot
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EvoBot
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: JSON
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: User login
      tags:
      - auth
...
  /kf:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all kf
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List KF
      tags:
      - kf
    post:
      consumes:
      - application/json
      description: Add a new kf
      parameters:
      - description: kf Information
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.KF'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Add kf
      tags:
      - kf
  /kf/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete the kf by UUID
      parameters:
      - description: kf UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete KF
      tags:
      - kf
    get:
      consumes:
      - application/json
      description: Get the kf information by UUID
      parameters:
      - description: kf UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get KF
      tags:
      - kf
    patch:
      consumes:
      - application/json
      description: Update the kf by UUID
      parameters:
      - description: kf UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: kf Information
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.KF'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update kf
      tags:
      - kf
  /llmapp/config:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all llmapp configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LLMAppConfig'
      summary: List llmapp configuration
      tags:
      - llmapp_config
    post:
      consumes:
      - application/json
      description: Add a new llmapp configuration
      parameters:
      - description: llmapp configuration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LLMAppConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Add new llmapp configuration
      tags:
      - llmapp
  /llmapp/config/{uuid}:
    get:
      consumes:
      - application/json
      description: Retrieve llmapp configuration by its UUID
      parameters:
      - description: UUID of the llmapp configuration
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get llmapp configuration by UUID
      tags:
      - llmapp
    patch:
      consumes:
      - application/json
      description: Update the llmapp configuration with the provided data
      parameters:
      - description: UUID of the llmapp configuration
        in: path
        name: uuid
        required: true
        type: string
      - description: llmapp configuration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LLMAppConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update llmapp configuration
      tags:
      - llmapp_config
  /policy:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all policy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List policy
      tags:
      - policy
    post:
      consumes:
      - application/json
      description: Add a new policy
      parameters:
      - description: policy Information
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.Policy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Add policy
      tags:
      - policy
  /policy/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete the policy by UUID
      parameters:
      - description: policy UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete policy
      tags:
      - policy
    get:
      consumes:
      - application/json
      description: Get the policy information by UUID
      parameters:
      - description: policy UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get policy
      tags:
      - policy
    patch:
      consumes:
      - application/json
      description: Update the policy by UUID
      parameters:
      - description: policy UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: policy Information
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.Policy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update policy
      tags:
      - policy
...
swagger: "2.0"